Refactor: Rename sensor reading API and add new endpoints
This commit renames `ApiSensorReading` and `ApiSensorReadingImpl` to `SensorAPI` and `SensorAPIImpl` respectively.

It also introduces new methods to the `SensorAPI` interface and its implementation for controlling various aspects of the sensor environment, including:
- Toggling light, fan, extractor, and pump
- Managing EC (Electrical Conductivity) levels (get, up, down)
- Managing pH levels (get, up, down)

The implementation for toggling the light and fan has been added, while other new methods are currently marked as `TODO`.
The `SensorAPIImpl` constructor has been updated to optionally accept an `isLocal` boolean to determine the base URL.
The Koin module `SensorReadingModule` has been updated to reflect the interface and class renaming.