Refactor Konnection instance creation
This commit refactors the creation of the Konnection instance.

Previously, Konnection.createInstance() was called in `MainApplication.kt` for Android and in the common `networkConfigModule`.

Now, `Konnection.createInstance()` is moved from the common module and `MainApplication.kt` to the platform-specific Koin modules for JVM (`Platform.jvm.kt`) and iOS (`Platform.ios.kt`). This ensures that Konnection is initialized appropriately for each platform within its respective dependency injection setup.