Refactor network connection logic and enable cleartext traffic
This commit refactors the `GetNetworkConnectionUseCase` to simplify the determination of local versus remote connections. It now directly uses the `currentIPV4Address` to decide the base URL and whether the connection is local.

Additionally, it enables cleartext HTTP traffic for the local IP address `192.168.8.14` in the Android manifest by adding a `network_security_config.xml` file.

The `createHttpClient` function is now provided as a singleton in both Android and iOS Koin modules.